srcs/files/backend/requirements.txt
Django==5.0.2
python-dotenv
termcolor
psycopg2-binary==2.9.9
djangorestframework
django-cors-headers